mdss: mdp: increase size of mdp image data length
MDP image structure's len member is currently 32-bit.
This field is incorrectly cast to unsigned long before
passing to ion APIs like ion_map_iommu that expect 64-bit
wide data. This can cause unexpected data overwrites.
Avoid such incorrect casts and declare the member
in mdp image structure to have 64-bit width.
Additionally, use appropriate format specifiers in print
statements for the new field type.

+++ drivers/video/msm/mdss/mdss_mdp.h
@@ -286,7 +286,7 @@ struct mdss_mdp_plane_sizes {
 
 struct mdss_mdp_img_data {
 	dma_addr_t addr;
-	u32 len;
+	unsigned long len;
 	u32 flags;
 	int p_need;
 	bool mapped;
